Fix for fail finding grep in C:\Git\bin\grep.exe
First of all, thanks for your work!

I recently updated counsel-etags. This update was causing counsel-etags-grep to fail with message `/usr/bin/bash.exe: nil Command not found`. So it seemed that for some reason counsel-etags was not finding a suitable grep/ag/rg. This was odd, because prior to the update it was found, and nothing else changed in my system.

I inspected this further and traced it back to (counsel-etags-guess-program), which in windows was looking for the executable only in Cygwin folders. I am using git's grep for windows and I don't even have Cygwin.

So I changed (counsel-etags-guess-program) to look first in the system path, and do the Cygwin search only if this fails. Don't know if this is correct, but at least fixes my problem. HTH.